+++ device-api/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [[_TOC_]]
 ## [Unreleased]
 - [SW-19030] Add commands to set/get vmin lut to/from etsoc persisted storage 
+- [SW-19030] Reuse asset buffer for fru and vmin lut commands 
 ### Added
 ### Changed
 ### Deprecated